CH GDP Growth Contributions

A list of time series containing sector contributions to Swiss GDP over time.

Usage

CHGDP

Format

List list of six time series of class ts, containing contributions to Swiss GDP growth

manufacturing

Growth contribution of manufacturing.

energy

Growth contribution of energy, water sector

construction

Growth contribution construction sector.

hotels

Growth contribution of hotels.

fin_insur

Growth contribution of financial services and insurances.

other

Growth contribution of other sectors.

Source

https://www.seco.admin.ch/seco/en/home/wirtschaftslage---wirtschaftspolitik/Wirtschaftslage/bip-quartalsschaetzungen-/daten.html