Deprecated function(s) in tstools

These functions are provided for compatibility with older version of the tstools package. They may eventually be completely removed.

Arguments

...

Parameters to be passed to the modern version of the function

Details

computeDecimalTime	now a synonym for compute_decimal_time

concatTs	now a synonym for concat_ts

fillupYearWitnNAs	now a synonym for fill_year_with_nas

importTimeSeries	now a synonym for read_ts

init_tsplot_theme	now a synonym for init_tsplot_theme

overlapSortedLists	now a synonym for overlap_sorted_ts_lists

overlapTslByName	now a synonym for overlap_ts_lists_by_name

resolveOverlap	now a synonym for resolve_ts_overlap

stripLeadingNAsFromTs	now a synonym for strip_ts_of_leading_nas

stripTrailingNAsFromTs	now a synonym for strip_ts_of_trailing_nas

writeTimeSeries	now a synonym for write_ts